requirements.txt
Pillow==9.5.0
pytesseract==0.3.10
pdf2image==1.16.3